Sex_SNP_finder v0.1.0
Sex_SNP_finder_Fst_Dxy.pl was updated to Sex_SNP_finder_GA.pl with the
GA standing for Genomics Analysis. Several key features were added.
First, a minimum allele count feature was added in order to try to
filter out sequencing errors. Setting this to 0 or 1 will therefore not
filter out any low frequency SNPs. Additionally, a maximum read depth
feature was added to exclude sites with high coverage and also allow
for an upper bound on Nei's D to be calculated. Calculations for Da, Cp
and Nei's D have been added. Formulas for these equations can be found
on the GitHub formula section. Nei's D for fixed sites is infinity,
which isn't terrible helpful when it is trying to be viewed in a genome
browser such as IGV. So, a "maximum Nei's D value" is calculated based
upon one population being fixed at the maximum read depth and the other
population being almost alternatively fixed for the opposite allele but
it has one count in common with the former population. For example if
the maximum_read_depth is equal to 100 then the "maximum Nei's D value"
would be calculated for population 1 being fixed for 100 Adenine counts
and population 2 would be polymorphic for 99 Thymine counts and 1
Adenine count. In this example the "maximum Nei's D value" is equal to
4.59517.